RR-320 - Support navigating to Add A Goal directly from CIAG
