+++ server/routes/overview/index.ts
@@ -2,13 +2,12 @@ import { Router } from 'express'
 import { Services } from '../../services'
 import { checkUserHasViewAuthority } from '../../middleware/roleBasedAccessControl'
 import OverviewController from './overviewController'
-import PrisonerSummaryRequestHandler from './prisonerSummaryRequestHandler'
+import { retrievePrisonerSummaryIfNotInSession } from '../routerRequestHandlers'
 
 /**
  * Route definitions for the pages relating to the main Overview page
  */
 export default (router: Router, services: Services) => {
-  const prisonerSummaryRequestHandler = new PrisonerSummaryRequestHandler(services.prisonerSearchService)
   const overViewController = new OverviewController(
     services.curiousService,
     services.educationAndWorkPlanService,
@@ -17,7 +16,7 @@ export default (router: Router, services: Services) => {
 
   router.use('/plan/:prisonNumber/view/*', [
     checkUserHasViewAuthority(),
-    prisonerSummaryRequestHandler.getPrisonerSummary,
+    retrievePrisonerSummaryIfNotInSession(services.prisonerSearchService),
   ])
 
   router.get('/plan/:prisonNumber/view/overview', [overViewController.getOverviewView])
