+++ server/routes/routerRequestHandlers.test.ts
@@ -1,21 +1,26 @@
 import { NextFunction, Response, Request } from 'express'
 import { SessionData } from 'express-session'
+import type { Prisoner } from 'prisonRegisterApiClient'
 import type { UpdateGoalForm } from 'forms'
 import type { NewGoal } from 'compositeForms'
+import createError from 'http-errors'
 import {
   checkCreateGoalFormExistsInSession,
   checkAddStepFormsArrayExistsInSession,
   checkPrisonerSummaryExistsInSession,
   checkUpdateGoalFormExistsInSession,
   checkNewGoalsFormExistsInSession,
+  retrievePrisonerSummaryIfNotInSession,
 } from './routerRequestHandlers'
 import { aValidAddStepForm } from '../testsupport/addStepFormTestDataBuilder'
 import aValidPrisonerSummary from '../testsupport/prisonerSummaryTestDataBuilder'
 import aValidCreateGoalForm from '../testsupport/createGoalFormTestDataBuilder'
 import aValidAddNoteForm from '../testsupport/addNoteFormTestDataBuilder'
+import { PrisonerSearchService } from '../services'
 
 describe('routerRequestHandlers', () => {
   const req = {
+    user: {} as Express.User,
     session: {} as SessionData,
     params: {} as Record<string, string>,
   }
@@ -26,6 +31,7 @@ describe('routerRequestHandlers', () => {
 
   beforeEach(() => {
     jest.resetAllMocks()
+    req.user = {} as Express.User
     req.session = {} as SessionData
     req.params = {} as Record<string, string>
   })
@@ -369,6 +375,146 @@ describe('routerRequestHandlers', () => {
       // Then
       expect(res.redirect).toHaveBeenCalledWith(`/plan/${prisonNumber}/goals/add-note`)
       expect(next).not.toHaveBeenCalled()
+    })
+  })
+
+  describe('retrievePrisonerSummaryIfNotInSession', () => {
+    const prisonerSearchService = {
+      getPrisonerByPrisonNumber: jest.fn(),
+    }
+
+    const requestHandler = retrievePrisonerSummaryIfNotInSession(
+      prisonerSearchService as unknown as PrisonerSearchService,
+    )
+
+    it('should retrieve prisoner and store in session given prisoner not in session', async () => {
+      // Given
+      const username = 'a-dps-user'
+      req.user.username = username
+
+      req.session.prisonerSummary = undefined
+
+      const prisonNumber = 'A1234GC'
+      const prisonId = 'MDI'
+      req.params.prisonNumber = prisonNumber
+      const prisoner = {
+        prisonerNumber: prisonNumber,
+        prisonId,
+        releaseDate: '2025-12-31',
+        firstName: 'Jimmy',
+        lastName: 'Lightfingers',
+        dateOfBirth: '1969-02-12',
+        receptionDate: '1999-08-29',
+      } as Prisoner
+      prisonerSearchService.getPrisonerByPrisonNumber.mockResolvedValue(prisoner)
+
+      const expectedPrisonerSummary = aValidPrisonerSummary(prisonNumber, prisonId)
+
+      // When
+      await requestHandler(req as undefined as Request, res as undefined as Response, next as undefined as NextFunction)
+
+      // Then
+      expect(prisonerSearchService.getPrisonerByPrisonNumber).toHaveBeenCalledWith(prisonNumber, username)
+      expect(req.session.prisonerSummary).toEqual(expectedPrisonerSummary)
+      expect(next).toHaveBeenCalled()
+    })
+
+    it('should retrieve prisoner and store in session given different prisoner already in session', async () => {
+      // Given
+      const username = 'a-dps-user'
+      req.user.username = username
+
+      req.session.prisonerSummary = aValidPrisonerSummary('Z1234XY', 'BXI')
+
+      const prisonNumber = 'A1234GC'
+      const prisonId = 'MDI'
+      req.params.prisonNumber = prisonNumber
+      const prisoner = {
+        prisonerNumber: prisonNumber,
+        prisonId,
+        releaseDate: '2025-12-31',
+        firstName: 'Jimmy',
+        lastName: 'Lightfingers',
+        dateOfBirth: '1969-02-12',
+        receptionDate: '1999-08-29',
+      } as Prisoner
+      prisonerSearchService.getPrisonerByPrisonNumber.mockResolvedValue(prisoner)
+
+      const expectedPrisonerSummary = aValidPrisonerSummary(prisonNumber, prisonId)
+
+      // When
+      await requestHandler(req as undefined as Request, res as undefined as Response, next as undefined as NextFunction)
+
+      // Then
+      expect(prisonerSearchService.getPrisonerByPrisonNumber).toHaveBeenCalledWith(prisonNumber, username)
+      expect(req.session.prisonerSummary).toEqual(expectedPrisonerSummary)
+      expect(next).toHaveBeenCalled()
+    })
+
+    it('should not retrieve prisoner given prisoner already in session', async () => {
+      // Given
+      const username = 'a-dps-user'
+      req.user.username = username
+
+      const prisonNumber = 'A1234GC'
+      const prisonId = 'MDI'
+
+      req.session.prisonerSummary = aValidPrisonerSummary(prisonNumber, prisonId)
+
+      req.params.prisonNumber = prisonNumber
+      const expectedPrisonerSummary = aValidPrisonerSummary(prisonNumber, prisonId)
+
+      // When
+      await requestHandler(req as undefined as Request, res as undefined as Response, next as undefined as NextFunction)
+
+      // Then
+      expect(prisonerSearchService.getPrisonerByPrisonNumber).not.toHaveBeenCalled()
+      expect(req.session.prisonerSummary).toEqual(expectedPrisonerSummary)
+      expect(next).toHaveBeenCalled()
+    })
+
+    it('should call next function with error given retrieving prisoner fails with a 404', async () => {
+      // Given
+      const username = 'a-dps-user'
+      req.user.username = username
+
+      req.session.prisonerSummary = undefined
+
+      const prisonNumber = 'A1234GC'
+      req.params.prisonNumber = prisonNumber
+
+      prisonerSearchService.getPrisonerByPrisonNumber.mockRejectedValue(createError(404, 'Not Found'))
+      const expectedError = createError(404, `Prisoner ${prisonNumber} not returned by the Prisoner Search Service API`)
+
+      // When
+      await requestHandler(req as undefined as Request, res as undefined as Response, next as undefined as NextFunction)
+
+      // Then
+      expect(prisonerSearchService.getPrisonerByPrisonNumber).toHaveBeenCalledWith(prisonNumber, username)
+      expect(req.session.prisonerSummary).toBeUndefined()
+      expect(next).toHaveBeenCalledWith(expectedError)
+    })
+
+    it('should call next function with error given retrieving prisoner fails with a 500', async () => {
+      // Given
+      const username = 'a-dps-user'
+      req.user.username = username
+
+      req.session.prisonerSummary = undefined
+
+      const prisonNumber = 'A1234GC'
+      req.params.prisonNumber = prisonNumber
+
+      prisonerSearchService.getPrisonerByPrisonNumber.mockRejectedValue(createError(500, 'Not Found'))
+      const expectedError = createError(500, `Prisoner ${prisonNumber} not returned by the Prisoner Search Service API`)
+
+      // When
+      await requestHandler(req as undefined as Request, res as undefined as Response, next as undefined as NextFunction)
+
+      // Then
+      expect(prisonerSearchService.getPrisonerByPrisonNumber).toHaveBeenCalledWith(prisonNumber, username)
+      expect(req.session.prisonerSummary).toBeUndefined()
+      expect(next).toHaveBeenCalledWith(expectedError)
     })
   })
 })
